Implement `stack_v3` as hypothetical substitute for `stack`, using new `head_info()`.
Also, refactor `stacks()` related code into its own module to shrink `lib.rs`.
